ISSUE #5299 - move pendingInitialFilters action and selector
