+++ frontend/src/v5/store/tickets/card/ticketsCard.selectors.ts
@@ -152,6 +152,11 @@ export const selectIsShowingPins = createSelector(
 	selectTicketsCardDomain, (state) => state.isShowingPins,
 );
 
+export const selectAreInitialFiltersPending = createSelector(
+	selectTicketsCardDomain,
+	(state) => state.areInitialFiltersPending,
+);
+
 export const selectCurrentModelFilteredTickets = createSelector(
 	(state) => state,
 	selectCurrentModel,
