+++ frontend/src/v5/store/tickets/tickets.redux.ts
@@ -46,7 +46,6 @@ export const { Types: TicketsTypes, Creators: TicketsActions } = createActions({
 	updateTicketGroup: ['teamspace', 'projectId', 'modelId', 'ticketId', 'group', 'isFederation'],
 	updateTicketGroupSuccess: ['group'],
 	clearGroups: [],
-	setAreInitialFiltersPending: ['isPending'],
 }, { prefix: 'TICKETS/' }) as { Types: Constants<ITicketsActionCreators>; Creators: ITicketsActionCreators };
 
 export const INITIAL_STATE: ITicketsState = {
@@ -54,7 +53,6 @@ export const INITIAL_STATE: ITicketsState = {
 	templatesByModelId: {},
 	groupsByGroupId: {},
 	riskCategories: [],
-	areInitialFiltersPending: true,
 };
 
 export const fetchTicketsSuccess = (state: ITicketsState, { modelId, tickets }: FetchTicketsSuccessAction) => {
@@ -110,11 +108,6 @@ export const clearGroups = (state: ITicketsState) => {
 	state.groupsByGroupId = {};
 };
 
-export const setAreInitialFiltersPending = (state: ITicketsState, { isPending }: SetAreInitialFiltersPendingAction) => {
-	console.log('@@ REDUX set pending', isPending);
-	state.areInitialFiltersPending = isPending;
-};
-
 export const ticketsReducer = createReducer(INITIAL_STATE, produceAll({
 	[TicketsTypes.FETCH_TICKETS_SUCCESS]: fetchTicketsSuccess,
 	[TicketsTypes.FETCH_TEMPLATES_SUCCESS]: fetchTemplatesSuccess,
@@ -124,7 +117,6 @@ export const ticketsReducer = createReducer(INITIAL_STATE, produceAll({
 	[TicketsTypes.FETCH_TICKET_GROUPS_SUCCESS]: fetchTicketGroupsSuccess,
 	[TicketsTypes.UPDATE_TICKET_GROUP_SUCCESS]: updateTicketGroupSuccess,
 	[TicketsTypes.CLEAR_GROUPS]: clearGroups,
-	[TicketsTypes.SET_ARE_INITIAL_FILTERS_PENDING]: setAreInitialFiltersPending,
 }));
 
 export interface ITicketsState {
@@ -132,7 +124,6 @@ export interface ITicketsState {
 	templatesByModelId: Record<string, ITemplate[]>,
 	riskCategories: string[],
 	groupsByGroupId: Record<string, Group>,
-	areInitialFiltersPending: boolean,
 }
 
 export type FetchTicketsAction = Action<'FETCH_TICKETS'> & TeamspaceProjectAndModel & { isFederation: boolean, propertiesToInclude?: string[] };
@@ -153,7 +144,6 @@ export type FetchTicketGroupsSuccessAction = Action<'FETCH_TICKET_GROUPS_SUCCESS
 export type UpdateTicketGroupAction = Action<'UPDATE_TICKET_GROUP'> & TeamspaceProjectAndModel & { ticketId: string, group: Group, isFederation: boolean };
 export type UpdateTicketGroupSuccessAction = Action<'UPDATE_TICKET_GROUP_SUCCESS'> & { group: Group };
 export type ClearGroupsAction = Action<'CLEAR_GROUPS'>;
-export type SetAreInitialFiltersPendingAction = Action<'SET_ARE_INITIALL_FILTERS_PENDING'> & { isPending: boolean };
 
 export interface ITicketsActionCreators {
 	fetchTickets: (
@@ -241,5 +231,4 @@ export interface ITicketsActionCreators {
 		group: Group,
 	) => UpdateTicketGroupSuccessAction;
 	clearGroups: () => ClearGroupsAction;
-	setAreInitialFiltersPending: (isPending: boolean) => SetAreInitialFiltersPendingAction,
 }
