+++ frontend/src/v5/store/tickets/tickets.selectors.ts
@@ -141,8 +141,3 @@ export const selectStatusConfigByTemplateId = createSelector(
 	(state, ...args) => selectCurrentProjectTemplateById(state, args.at(-1)),
 	(ticketTemplate, projectTemplate) => ticketTemplate?.config?.status || projectTemplate?.config?.status || DEFAULT_STATUS_CONFIG,
 );
-
-export const selectAreInitialFiltersPending = createSelector(
-	selectTicketsDomain,
-	(state) => state.areInitialFiltersPending,
-);
